+++ src/js/FormModel.js
@@ -722,7 +722,6 @@ define( [ 'xpath', 'jquery', 'enketo-js/plugins', 'enketo-js/extend', 'jquery.xp
         return expr;
     };
 
-
     FormModel.prototype.nsResolver = {
 
         lookupNamespaceURI: function( prefix ) {
@@ -780,6 +779,47 @@ define( [ 'xpath', 'jquery', 'enketo-js/plugins', 'enketo-js/extend', 'jquery.xp
     };
 
     /**
+     * Replaces indexed-repeat(node, path, position, path, position, etc) substrings by converting them
+     * to their native XPath equivalents using [position() = x] predicates
+     *
+     * @param  {string} expr the XPath expression
+     * @return {string}      converted XPath expression
+     */
+    FormModel.prototype.replaceIndexedRepeatFn = function( expr ) {
+        var indexedRepeats = expr.match( /(indexed-repeat\s?\([^\)]+\))/g );
+
+        if ( !indexedRepeats ) {
+            return expr;
+        }
+
+        indexedRepeats.forEach( function( indexedRepeat, irIndex ) {
+            var i, positionedPath,
+                params = indexedRepeat.replace( /indexed-repeat\s?\(([^\)]+)\)/g, '$1' ).split( ',' );
+
+            if ( params.length % 2 === 1 ) {
+
+                //trim parameters
+                params.forEach( function( param, index ) {
+                    params[ index ] = param.trim();
+                } );
+
+                positionedPath = params[ 0 ];
+
+                for ( i = params.length - 1; i > 1; i -= 2 ) {
+                    positionedPath = positionedPath.replace( params[ i - 1 ], params[ i - 1 ] + '[position() = ' + params[ i ] + ']' );
+                }
+                expr = expr.replace( indexedRepeat, positionedPath );
+
+            } else {
+                console.error( 'indexed repeat with incorrect number of parameters found', indexedRepeat );
+                return "'Error with indexed-repeat parameters'";
+            }
+        } );
+
+        return expr;
+    };
+
+    /**
      * Evaluates an XPath Expression using XPathJS_javarosa (not native XPath 1.0 evaluator)
      *
      * This function does not seem to work properly for nodeset resulttypes otherwise:
@@ -808,6 +848,7 @@ define( [ 'xpath', 'jquery', 'enketo-js/plugins', 'enketo-js/extend', 'jquery.xp
         expr = this.shiftRoot( expr );
         expr = this.replaceInstanceFn( expr );
         expr = this.replaceCurrentFn( expr );
+        expr = this.replaceIndexedRepeatFn( expr );
 
         // path corrections for repeated nodes: http://opendatakit.github.io/odk-xform-spec/#a-big-deviation-with-xforms
         if ( selector ) {